Shifted classes, integrated setting retrieval and persistence in new config form, UI work
