+++ Mamesaver.Config/ConfigForm.xaml
@@ -5,6 +5,7 @@
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:config="clr-namespace:Mamesaver.Config"
     xmlns:mvvm="http://prismlibrary.com/"
+    xmlns:converters="clr-namespace:Mamesaver.Config.Converters"
     mvvm:ViewModelLocator.AutoWireViewModel="True"
     x:Class="Mamesaver.Config.ConfigForm"
         mc:Ignorable="d"
@@ -12,11 +13,12 @@
     ResizeMode="NoResize" SizeToContent="WidthAndHeight">
 
     <Window.Resources>
-        <ResourceDictionary Source="Styles.xaml"/>
-   </Window.Resources>
+        <converters:InvertedBooleanConverter x:Key="InvertedBooleanConverter"/>
+        <ResourceDictionary x:Key="Styles" Source="Styles.xaml"/>
+    </Window.Resources>
 
     <DockPanel LastChildFill="False" Margin="10">
-        <TabControl DockPanel.Dock="Top" HorizontalAlignment="Left" Margin="0,0,0,10" HorizontalContentAlignment="Stretch" Width="800" Height="600">
+        <TabControl DockPanel.Dock="Top" HorizontalAlignment="Left" Margin="0,0,0,10" HorizontalContentAlignment="Stretch" Width="880" Height="600">
             <TabItem Header="General">
                 <config:GeneralTab/>
             </TabItem>
@@ -39,8 +41,8 @@
         <Button Command="{Binding ResetToDefaultsClick}" DockPanel.Dock="Left" Width="120" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="0,5,5,0">_Reset to Defaults</Button>
 
         <StackPanel Orientation="Horizontal" DockPanel.Dock="Right">
-            <Button Command="{Binding OkClick}" Width="90" IsCancel="True" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="5,5,5,0">_OK</Button>
-            <Button Command="{Binding CancelClick}" Width="90" IsDefault="True" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="5,5,0,0">_Cancel</Button>
+            <Button Command="{Binding OkClick}" Width="90" IsDefault="True" IsEnabled="{Binding Path=Rebuilding, Converter={StaticResource InvertedBooleanConverter}}" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="5,5,5,0">_OK</Button>
+            <Button Command="{Binding CancelClick}" Width="90" IsCancel="True" HorizontalAlignment="Left" VerticalAlignment="Bottom" Margin="5,5,0,0">_Cancel</Button>
         </StackPanel>
     </DockPanel>
 </Window>
