+++ Mamesaver.Config/Converters/InvertedBooleanConverter.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Globalization;
+using System.Windows.Data;
+
+namespace Mamesaver.Config.Converters
+{
+    [ValueConversion(typeof(bool), typeof(bool))]
+    public class InvertedBooleanConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (targetType != typeof(bool)) throw new InvalidOperationException("The target must be a boolean");
+            return !(bool?) value ?? false;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotSupportedException();
+    }
+}
