+++ Mamesaver.Config/Converters/InvertedBooleanToVisibilityConverter.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Globalization;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+
+namespace Mamesaver.Config.Converters
+{
+    [ValueConversion(typeof(bool), typeof(Visibility))]
+    public class InvertedBooleanToVisibilityConverter : IValueConverter
+    {
+        private readonly BooleanToVisibilityConverter _converter = new BooleanToVisibilityConverter();
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            var visibility = (Visibility)_converter.Convert(value, targetType, parameter, culture);
+            return visibility == Visibility.Collapsed ? Visibility.Visible : Visibility.Collapsed;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) => throw new NotSupportedException();
+    }
+}
