+++ Mamesaver.Config/Models/GameViewModel.cs
@@ -1,35 +1,12 @@
-﻿/**
- * Licensed under The MIT License
- * Redistributions of files must retain the above copyright notice.
- */
-
-using System;
-using System.ComponentModel;
+﻿using System.ComponentModel;
 using System.Runtime.CompilerServices;
-using System.Xml.Serialization;
 using JetBrains.Annotations;
 
 namespace Mamesaver.Config.Models
 {
-    [Serializable, XmlRoot("game")]
-    public class Game  : INotifyPropertyChanged
+    public class GameViewModel  : INotifyPropertyChanged
     {
         private bool _selected = true;
-        public string Category { get; set; }
-
-        public Game()
-        {
-        }
-
-        public Game(string name, string description, string year, string manufacturer, string rotation, string category = null)
-        {
-            Category = category;
-            Name = name;
-            Description = description;
-            Year = year;
-            Manufacturer = manufacturer;
-            Rotation = rotation;
-        }
 
         public bool Selected
         {
@@ -42,20 +19,18 @@ namespace Mamesaver.Config.Models
             }
         }
 
-        [XmlAttribute("name")]
         public string Name { get; set; }
-
-        [XmlAttribute("description")]
         public string Description { get; set; }
-
-        [XmlAttribute("year")]
         public string Year { get; set; }
 
-        [XmlAttribute("manufacturer")]
         public string Manufacturer { get; set; }
-
-        [XmlAttribute("rotation")]
         public string Rotation { get; set; }
+        public string Category { get; set; }
+
+        public string ManufacturerSort => $"{Manufacturer}:{Description}";
+        public string CategorySort => $"{Category}:{Description}";
+        public string YearSort => $"{Year}:{Description}";
+        public string RotationSort => $"{Rotation}:{Description}";
 
         public event PropertyChangedEventHandler PropertyChanged;
