+++ Mamesaver.Config/ViewModels/ConfigFormViewModel.cs
@@ -1,42 +1,38 @@
-﻿using System.Collections.Generic;
-using System.Collections.ObjectModel;
+﻿using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Input;
 using JetBrains.Annotations;
+using Mamesaver.Config.Models;
 using Mamesaver.Models;
 using Mamesaver.Models.Configuration;
 using Mamesaver.Services.Categories;
 using Mamesaver.Services.Configuration;
+using Mamesaver.Services.Mame;
 using Prism.Commands;
+using Serilog;
 
 namespace Mamesaver.Config.ViewModels
 {
     public class ConfigFormViewModel : INotifyPropertyChanged
     {
-        private readonly Settings _settings;
-        private readonly LayoutSettings _layoutSettings;
-        private readonly AdvancedSettings _advancedSettings;
+        private Settings _settings;
+        private LayoutSettings _layoutSettings;
+        private AdvancedSettings _advancedSettings;
         private readonly GameList _gameList;
         private readonly GeneralSettingsStore _generalSettingsStore;
         private readonly GameListStore _gameListStore;
-        //private readonly GameListBuilder _gameListBuilder;
-        private const string DefaultFont = "Arial";
-        private const int DefaultFontSize = 13;
+        private readonly GameListBuilder _gameListBuilder;
 
         private bool? _allSelected = true;
-        private bool _cloneToAllMonitors;
-        private string _commandLineOptions;
-        private IEnumerable<SelectableGame> _filteredGames;
-        private bool _inGameTitlesEnabled;
-        private string _mamePath;
-        private bool _splashEnabled;
-        private bool _includeImperfectEmulation;
-        private bool _debugLogging;
+        private IEnumerable<GameViewModel> _filteredGames;
 
         public ConfigFormViewModel(
             Settings settings,
@@ -44,8 +40,8 @@ namespace Mamesaver.Config.ViewModels
             AdvancedSettings advancedSettings,
             GameList gameList,
             GeneralSettingsStore generalSettingsStore,
-            GameListStore gameListStore)
-            //GameListBuilder gameListBuilder)
+            GameListStore gameListStore,
+            GameListBuilder gameListBuilder)
         {
             _settings = settings;
             _layoutSettings = layoutSettings;
@@ -53,13 +49,14 @@ namespace Mamesaver.Config.ViewModels
             _gameList = gameList;
             _generalSettingsStore = generalSettingsStore;
             _gameListStore = gameListStore;
-            //_gameListBuilder = gameListBuilder;
+            _gameListBuilder = gameListBuilder;
 
+            // TODO shift out of constructor plz
             LoadFonts();
             LoadGames();
         }
 
-        public List<SelectableGame> Games { get; set; }
+        public List<GameViewModel> Games { get; set; }
 
         public bool? AllSelected
         {
@@ -83,6 +80,9 @@ namespace Mamesaver.Config.ViewModels
 
         private void SaveAndClose()
         {
+            _generalSettingsStore.Save();
+            _gameListStore.Save();
+
             Close();
         }
 
@@ -93,11 +93,18 @@ namespace Mamesaver.Config.ViewModels
 
         private void ResetToDefaults()
         {
+            // Preserve MAME executable path
+            var executablePath = _settings.ExecutablePath;
+            _settings = new Settings { ExecutablePath = executablePath };
 
+            _layoutSettings = _settings.LayoutSettings;
+            _advancedSettings = _settings.AdvancedSettings;
+
+            // Notify all properties
+            OnPropertyChanged("");
         }
 
-
-        public IEnumerable<SelectableGame> FilteredGames
+        public IEnumerable<GameViewModel> FilteredGames
         {
             get => _filteredGames;
             set
@@ -111,35 +118,33 @@ namespace Mamesaver.Config.ViewModels
 
         public string GameCount => $"Num Games: {Games.Count}";
 
-        public string MamePath
+        public string ExecutablePath
         {
-            get => _mamePath;
+            get => _settings.ExecutablePath;
             set
             {
-                if (value == _mamePath) return;
-                _mamePath = value;
+                if (value == _settings.ExecutablePath) return;
+                _settings.ExecutablePath = value;
                 OnPropertyChanged();
             }
         }
-
         public string CommandLineOptions
         {
-            get => _commandLineOptions;
+            get => _settings.CommandLineOptions;
             set
             {
-                if (value == _commandLineOptions) return;
-                _commandLineOptions = value;
+                if (value == _settings.CommandLineOptions) return;
+                _settings.CommandLineOptions = value;
                 OnPropertyChanged();
             }
         }
-
-        public bool CloneToAllMonitors
+        public bool CloneScreen
         {
-            get => _cloneToAllMonitors;
+            get => _settings.CloneScreen;
             set
             {
-                if (value == _cloneToAllMonitors) return;
-                _cloneToAllMonitors = value;
+                if (value == _settings.CloneScreen) return;
+                _settings.CloneScreen = value;
                 OnPropertyChanged();
             }
         } 
@@ -149,54 +154,110 @@ namespace Mamesaver.Config.ViewModels
 
         public bool SplashEnabled
         {
-            get => _splashEnabled;
+            get => _layoutSettings.SplashScreen.Enabled;
             set
             {
-                if (value == _splashEnabled) return;
-                _splashEnabled = value;
+                if (value == _layoutSettings.SplashScreen.Enabled) return;
+                _layoutSettings.SplashScreen.Enabled = value;
                 OnPropertyChanged();
             }
         }
 
         public bool InGameTitlesEnabled
         {
-            get => _inGameTitlesEnabled;
+            get => _layoutSettings.InGameTitles.Enabled;
             set
             {
-                if (value == _inGameTitlesEnabled) return;
-                _inGameTitlesEnabled = value;
+                if (value == _layoutSettings.InGameTitles.Enabled) return;
+                _layoutSettings.InGameTitles.Enabled = value;
                 OnPropertyChanged();
             }
         }
 
-        public int SplashDuration { get; set; } = 3;
+        public int SplashDuration
+        {
+            get => _layoutSettings.SplashScreen.DurationSeconds;
+            set
+            {
+                if (value == _layoutSettings.SplashScreen.DurationSeconds) return;
+                _layoutSettings.SplashScreen.DurationSeconds = value;
+                OnPropertyChanged();
+            }
+        }
 
-        public string InGameFont { get; set; } = DefaultFont;
-        public int InGameFontSize { get; set; } = DefaultFontSize;
-        public string SplashFont { get; set; } = DefaultFont;
+        public string InGameFont
+        {
+            get => _layoutSettings.InGameTitles.FontSettings.Face;
+            set
+            {
+                if (value == _layoutSettings.InGameTitles.FontSettings.Face) return;
+                _layoutSettings.InGameTitles.FontSettings.Face = value;
+                OnPropertyChanged();
+            }
+        }
 
-        public bool HotkeysEnabled { get; set; }
+        public int InGameFontSize
+        {
+            get => _layoutSettings.InGameTitles.FontSettings.Size;
+            set
+            {
+                if (value == _layoutSettings.InGameTitles.FontSettings.Size) return;
+                _layoutSettings.InGameTitles.FontSettings.Size = value;
+                OnPropertyChanged();
+            }
+        }
 
-        public int Interval { get; set; } = 5;
+        public string SplashFont
+        {
+            get => _layoutSettings.InGameTitles.FontSettings.Face;
+            set
+            {
+                if (value == _layoutSettings.InGameTitles.FontSettings.Face) return;
+                _layoutSettings.InGameTitles.FontSettings.Face = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public bool HotkeysEnabled
+        {
+            get => _settings.HotKeys;
+            set
+            {
+                if (value == _settings.HotKeys) return;
+                _settings.HotKeys = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public int MinutesPerGame
+        {
+            get => _settings.MinutesPerGame;
+            set
+            {
+                if (value == _settings.MinutesPerGame) return;
+                _settings.MinutesPerGame = value;
+                OnPropertyChanged();
+            }
+        }
 
         public bool IncludeImperfectEmulation
         {
-            get => _includeImperfectEmulation;
+            get => _advancedSettings.IncludeImperfectEmulation;
             set
             {
-                if (value == _includeImperfectEmulation) return;
-                _includeImperfectEmulation = value;
+                if (value == _advancedSettings.IncludeImperfectEmulation) return;
+                _advancedSettings.IncludeImperfectEmulation = value;
                 OnPropertyChanged();
             }
         }
 
         public bool DebugLogging
         {
-            get => _debugLogging;
+            get => _advancedSettings.DebugLogging;
             set
             {
-                if (value == _debugLogging) return;
-                _debugLogging = value;
+                if (value == _advancedSettings.DebugLogging) return;
+                _advancedSettings.DebugLogging = value;
                 OnPropertyChanged();
             }
         }
@@ -227,7 +288,63 @@ namespace Mamesaver.Config.ViewModels
             else AllSelected = null;
         }
 
-        public void RebuildList() => LoadGames();
+        public bool Rebuilding
+        {
+            get => _rebuilding;
+            set
+            {
+                if (value == _rebuilding) return;
+                _rebuilding = value;
+                OnPropertyChanged();
+            }
+        }
+
+        private int _progress;
+        private bool _rebuilding;
+
+        public int Progress
+        {
+            get => _progress;
+            set
+            {
+                if (value == _progress) return;
+                _progress = value;
+                OnPropertyChanged();
+            }
+        }
+
+        public void RebuildList()
+        {
+            Progress = 0;
+            Rebuilding = true;
+            Games = new List<GameViewModel>();
+            OnPropertyChanged(nameof(GameCount));
+
+            Task.Run(() =>
+            {
+                try
+                {
+                    Games = _gameListBuilder
+                        .GetGameList(progress => Progress = progress)
+                        .Select(ToViewModel)
+                        .ToList();
+                }
+                catch (FileNotFoundException fe)
+                {
+                    MessageBox.Show(@"Error running MAME; verify that the executable path is correct.", @"Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                    Log.Error(fe, $"Unable to find MAME at {fe.FileName}");
+ 
+                }
+                catch (Exception ex)
+                {
+                    MessageBox.Show(@"Error running MAME; verify that the configuration is correct.", @"Error", MessageBoxButton.OK, MessageBoxImage.Error);
+                    Log.Error(ex, "Unable to construct game list");
+                }
+
+                Rebuilding = false;
+            });
+        }
+
         private void ClearFilters() => LoadGames();
 
         private void LoadFonts()
@@ -241,19 +358,46 @@ namespace Mamesaver.Config.ViewModels
 
         private void LoadGames()
         {
-            Games = _gameList.Games;
+            Games = _gameList.Games.Select(ToViewModel).ToList();
 
-            // TEMP performance testing
-            for(var i = 0; i < 5; i++) Games.AddRange(_gameList.Games);
+            //// TEMP performance testing
+            //for(var i = 0; i < 5; i++) Games.AddRange(_gameList.Games);
 
-            // Populate missing categories
+            //// Populate missing categories
             Games
                 .Where(game => game.Category == null)
                 .ToList()
                 .ForEach(game => game.Category = CategoryParser.GetCategory(game.Name));
 
             // Maintain a separate list of games displayed by the current filter
-            FilteredGames = new List<SelectableGame>(Games);
+            FilteredGames = new List<GameViewModel>(Games).ToList();
+        }
+
+        private GameViewModel ToViewModel(SelectableGame game)
+        {
+            var viewModel = new GameViewModel
+            {
+                Selected = game.Selected,
+                Name = game.Name,
+                Year = game.Year,
+                Description = game.Description,
+                Manufacturer = game.Manufacturer,
+                Category = game.Category,
+            };
+
+            switch (game.Rotation)
+            {
+                case "0":
+                case "180":
+                    viewModel.Rotation = "Horizontal";
+                    break;
+                case "90":
+                case "270":
+                    viewModel.Rotation = "Vertical";
+                    break;
+            }
+
+            return viewModel;
         }
 
         [NotifyPropertyChangedInvocator]
