+++ Mamesaver.Models/SelectableGame.cs
@@ -7,12 +7,12 @@ using System;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Xml.Serialization;
-using Mamesaver.Models.Annotations;
+using Mamesaver.Models.Properties;
 
 namespace Mamesaver.Models
 {
     [Serializable, XmlRoot("SelectableGame")]
-    public class SelectableGame : Game, INotifyPropertyChanged
+    public class SelectableGame : Game
     {
         private bool _selected;
 
@@ -26,7 +26,9 @@ namespace Mamesaver.Models
             Selected = selected;
         }
 
+    
         // FIXME this should be in a separate model
+        // OR find a way to notify this has changed
         [XmlAttribute("selected")]
         public bool Selected
         {
