+++ Mamesaver.Services/Mamesaver.Services.csproj
@@ -54,6 +54,9 @@
     <Compile Include="Configuration\GameListStore.cs" />
     <Compile Include="Configuration\GeneralSettingsStore.cs" />
     <Compile Include="Configuration\SettingsStore.cs" />
+    <Compile Include="Mame\GameListBuilder.cs" />
+    <Compile Include="Mame\MameInvoker.cs" />
+    <Compile Include="Mame\MamePathManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\Resource.Designer.cs">
       <AutoGen>True</AutoGen>
@@ -61,6 +64,14 @@
       <DependentUpon>Resource.resx</DependentUpon>
     </Compile>
     <Compile Include="ServiceResolver.cs" />
+    <Compile Include="Windows\MonitorInterop.cs" />
+    <Compile Include="Windows\PlatformInvokeGDI32.cs" />
+    <Compile Include="Windows\PlatformInvokeKernel32.cs" />
+    <Compile Include="Windows\PlatformInvokePowrprof.cs" />
+    <Compile Include="Windows\PlatformInvokeUSER32.cs" />
+    <Compile Include="Windows\PowerInterop.cs" />
+    <Compile Include="Windows\UserActivityHook.cs" />
+    <Compile Include="Windows\WindowsInterop.cs" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Serilog.Sinks.File">
