+++ Mamesaver.Services/Windows/PlatformInvokeKernel32.cs
@@ -1,7 +1,8 @@
 ﻿using System.Runtime.InteropServices;
+
 // ReSharper disable InconsistentNaming
 
-namespace Mamesaver.Windows
+namespace Mamesaver.Services.Windows
 {
     internal static class PlatformInvokeKernel32
     {
