+++ Mamesaver.Test/Integration/GameListStoreTests.cs
@@ -2,8 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using FluentAssertions;
-using Mamesaver.Configuration;
 using Mamesaver.Models;
+using Mamesaver.Services.Configuration;
 using Mamesaver.Test.Unit;
 using NUnit.Framework;
