+++ Mamesaver/GamePlayManager.cs
@@ -3,12 +3,14 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Threading;
 using System.Windows.Forms;
-using Mamesaver.Extensions;
 using Mamesaver.Hotkeys;
 using Mamesaver.Layout;
+using Mamesaver.Models;
 using Mamesaver.Models.Configuration;
+using Mamesaver.Models.Extensions;
 using Mamesaver.Power;
 using Mamesaver.Services.Configuration;
+using Mamesaver.Services.Mame;
 using Serilog;
 
 namespace Mamesaver
