+++ Mamesaver/Hotkeys/HotKeyManager.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Windows.Forms;
 using Mamesaver.Models.Configuration;
 using Mamesaver.Power;
-using Mamesaver.Windows;
+using Mamesaver.Services.Windows;
 using Serilog;
 
 namespace Mamesaver.Hotkeys
