--- Mamesaver/Windows/PlatformInvokeGDI32.cs
@@ -1,51 +1,0 @@
-﻿using System;
-using System.Runtime.InteropServices;
-// ReSharper disable InconsistentNaming
-
-namespace Mamesaver.Windows
-{
-    internal class PlatformInvokeGdi32
-    {
-        internal const int SRCOPY = 13369376;
-
-        [DllImport("gdi32.dll", EntryPoint = "DeleteDC")]
-        internal static extern IntPtr DeleteDC(IntPtr hDc);
-
-        [DllImport("gdi32.dll", EntryPoint = "DeleteObject")]
-        internal static extern IntPtr DeleteObject(IntPtr hDc);
-
-        [DllImport("gdi32.dll", EntryPoint = "BitBlt")]
-        internal static extern bool BitBlt(IntPtr hdcDest, int xDest, int yDest, int wDest, int hDest, IntPtr hdcSource, int xSrc, int ySrc, int rasterOp);
-
-        [DllImport("gdi32.dll", EntryPoint = "StretchBlt")]
-        internal static extern bool StretchBlt(IntPtr hdcDest, int xDest, int yDest, int wDest, int hDest, IntPtr hdcSource, int xSrc, int ySrc, int wSrc, int hSrc, int RasterOp);
-
-        [DllImport("gdi32.dll")]
-        internal static extern IntPtr CreateCompatibleDC(IntPtr hdc);
-
-        [DllImport("gdi32.dll")]
-        internal static extern IntPtr CreateCompatibleBitmap(IntPtr hdc, int nWidth, int nHeight);
-
-        [DllImport("gdi32.dll", EntryPoint = "SelectObject")]
-        internal static extern IntPtr SelectObject(IntPtr hdc, IntPtr bmp);
-
-        [DllImport("gdi32.dll", CharSet = CharSet.Auto, SetLastError = true, ExactSpelling = true)]
-        internal static extern int GetDeviceCaps(IntPtr hDC, int nIndex);
-
-        /// <summary>
-        /// DeviceCap - Device capabilities
-        /// </summary>
-        internal enum DeviceCap
-        {
-            /// <summary>
-            /// Logical pixels inch in X
-            /// </summary>
-            LOGPIXELSX = 88,
-            /// <summary>
-            /// Logical pixels inch in Y
-            /// </summary>
-            LOGPIXELSY = 90
-
-        }
-    }
-}
