--- Mamesaver/Windows/PlatformInvokeUSER32.cs
@@ -1,54 +1,0 @@
-﻿using System;
-using System.Runtime.InteropServices;
-// ReSharper disable InconsistentNaming
-
-namespace Mamesaver.Windows
-{
-    internal class PlatformInvokeUser32
-    {
-        internal const int SM_CXSCREEN = 0;
-        internal const int SM_CYSCREEN = 1;
-        internal const int SW_MINIMIZE = 6;
-
-        [DllImport("user32.dll", CharSet = CharSet.Auto)]
-        internal static extern IntPtr SendMessage(IntPtr hWnd, int Msg, int wParam, IntPtr lParam);
-
-        [DllImport("user32.dll")]
-        internal static extern void SetWindowPos(IntPtr hwnd, IntPtr hwndInsertAfter, int x, int y, int width, int height, uint flags);
-
-        [DllImport("user32.dll", EntryPoint = "GetDesktopWindow")]
-        internal static extern IntPtr GetDesktopWindow();
-
-        [DllImport("user32.dll", EntryPoint = "GetDC")]
-        internal static extern IntPtr GetDC(IntPtr ptr);
-
-        [DllImport("user32.dll", EntryPoint = "GetSystemMetrics")]
-        internal static extern int GetSystemMetrics(int abc);
-
-        [DllImport("user32.dll", EntryPoint = "GetWindowDC")]
-        internal static extern IntPtr GetWindowDC(Int32 ptr);
-
-        [DllImport("user32.dll", EntryPoint = "ReleaseDC")]
-        internal static extern IntPtr ReleaseDC(IntPtr hWnd, IntPtr hDC);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        internal static extern bool GetWindowRect(IntPtr hWnd, out RECT lpRect);
-
-        [DllImport("user32.dll")]
-        [return: MarshalAs(UnmanagedType.Bool)]
-        internal static extern bool ShowWindow(IntPtr hWnd, int nCmdShow);
-
-        [StructLayout(LayoutKind.Sequential)]
-        internal struct RECT
-        {
-            internal int Left;        // x position of upper-left corner
-            internal int Top;         // y position of upper-left corner
-            internal int Right;       // x position of lower-right corner
-            internal int Bottom;      // y position of lower-right corner
-        }
-
-        [DllImport("user32.dll")]
-        internal static extern bool SetProcessDPIAware();
-    }
-}
